darc-unarc: wire up l/v, and lift the listing out of the darc binary
Step 3. `unarc` now does l, v, t, x and e -- the whole non-SFX command set.

Two lifts, in the order the second needed the first:

  * `format_time`, `civil_from_days` and both `local_offset_seconds` (the Unix
    and Windows variants, with their `struct tm` and externs) move to a new
    `darc-arc/src/timefmt.rs`. This is the FFI-and-platform-split cluster that
    made me back out of lifting `list` in the previous step; on its own it is
    a clean move.
  * `list` then moves to `extract.rs`, beside `run_blocks`.

**The C unarc's listing disagrees with the archiver's, and the Rust one agrees.**
Measured on the same archive:

    C     1970-01-01 09:00:00       -dir- ./sub
          1970-01-01 09:00:00       20000 ./c.bin
    darc  1970-01-01 09:00:00       -dir- sub
          1970-01-01 09:00:00      20.000 c.bin

The C keeps the `./` that `arc a … .` stores, and never groups the digits.
`unarc l` is now byte-identical to `darc l`, which is the point of the port
rather than a side effect.

So `unarc-check.sh` gates the listing against `darc`, NOT against the C:
requiring the Rust to reproduce the C here would gate the drift in. Extraction
is still compared C-to-Rust, because there the two agree and should.

Teeth verified for the new check: pointed at a wrapper that answers `l` with the
C's listing, all nine cases report `listing`.

9 archives 0 differing, 696 tests, arc-golden-check 118/118, arc-cli-check 24/24
(it gates listing data too, so the lift is covered twice), clippy gate clean.
